requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
factor_analyzer==0.5.1
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==2.1.2
numdifftools==0.9.41
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
plotly==6.3.0
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
semopy==2.3.11
six==1.17.0
smmap==5.0.2
statsmodels==0.14.5
streamlit==1.48.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.5
